requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
Flask-Dance>=7.1.0,<8.0.0
